# service/liteaccessservice/swagger-ui/swagger.yaml
swagger: "2.0"
info:
  description: >-
    A REST interface for state queries, transaction generation and
    send.
  version: "1.0.2.rc.1"
  title: ankrchain-las Swagger-UI
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
schemes:
- "http"
- "https"
paths:
...
  /v1/account/generate/{count}:
    get:
      tags:
      - Account
      summary: Generate new ankrchain accounts
      description: >-
        Generate new ankrchain accounts
      operationId: "generateAccounts"
      produces:
      - "application/json"
      parameters:
      - name: "count"
        in: "path"
        description: "The account count to generate"
        required: false
        type: number
      responses:
        '200':
          description: Sucessfully generate new ankrchain accounts
          content:
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/definitions/AccountInfo"
        '400':
          description: The parameter was malformated
        '500':
          description: Server internal error
...
definitions:
...
  AccountInfo:
    type: object
    properties:
      priKey:
        type: string
      pubKey:
        type: string
      address:
        type: string
...